pyproject.toml
[project]
name = "ikos"
version = "0.1.0"
description = "Intelligent Knowledge Building System - From web information to structured knowledge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # 模型调用
    "ollama>=0.4.0",
    "openai>=1.0.0",
    # 模型下载（魔塔社区优先）
    "modelscope>=1.9.0",
    "huggingface_hub>=0.19.0",
    # 浏览器自动化与网页抓取
    "playwright>=1.40.0",
    # 向量数据库
    "chromadb>=0.4.0",
    "pymilvus>=2.3.0",
    # 图数据库
    "neo4j>=5.0.0",
    # 桌面 UI
    "pyqt6>=6.6.0",
    # 配置与数据验证
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    # 日志
    "loguru>=0.7.0",
    # 原生推理引擎
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "bitsandbytes>=0.41.0",
    # 硬件检测
    "nvidia-ml-py>=12.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    # lint + format（统一用 ruff）
    "ruff>=0.3.0",

    # 类型检查
    "pyright>=1.1.0",

    # 测试框架
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
ikos = "ikos.main:main"
ikos-cache = "ikos.utils.cache_cli:main"

[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ikos"]

# Ruff：统一负责格式化 + lint
[tool.ruff]
line-length = 100
target-version = "py313"
src = ["ikos"]
exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # 行长由 ruff format 管
]

[tool.ruff.lint.isort]
known-first-party = ["ikos"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Pyright 类型检查
[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
include = ["ikos"]
exclude = ["**/__pycache__", ".venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=ikos --cov-report=html"
